chore(api): move the sources API routes to the admin repository
