+++ api/v1/paths/admin/sources/{id}/imports.js
@@ -1,5 +1,5 @@
-const { processResponse } = require("../../utils");
-const configManager = require("../../../../../bin/configManager.");
+const { processResponse } = require("../../../utils");
+const configManager = require("../../../../../../bin/configManager.");
 module.exports = function () {
     let operations = {
         POST,
